# Jpeg Decoder: split decoder phases, refactor error handling, allow EOF in scans, fix #18