--- a/src/ImageSharp.Formats.Jpeg/Components/Block8x8F.cs
+++ b/src/ImageSharp.Formats.Jpeg/Components/Block8x8F.cs
@@ -6,6 +6,7 @@
 namespace ImageSharp.Formats.Jpg
 {
     using System;
+    using System.Buffers;
     using System.Numerics;
     using System.Runtime.CompilerServices;
     using System.Runtime.InteropServices;