--- a/src/ImageSharp.Formats.Jpeg/Components/Decoder/Bits.cs
+++ b/src/ImageSharp.Formats.Jpeg/Components/Decoder/Bits.cs
@@ -5,7 +5,6 @@
 
 namespace ImageSharp.Formats.Jpg
 {
-    using System;
     using System.Runtime.CompilerServices;
 
     /// <summary>
@@ -37,11 +36,11 @@ namespace ImageSharp.Formats.Jpg
         /// the caller is the one responsible for first checking that bits.UnreadBits &lt; n.
         /// </summary>
         /// <param name="n">The number of bits to ensure.</param>
-        /// <param name="decoder">Jpeg decoder</param>
+        /// <param name="bufferProcessor">The <see cref="BufferProcessor"/></param>
         [MethodImpl(MethodImplOptions.AggressiveInlining)]
-        public void EnsureNBits(int n, JpegDecoderCore decoder)
+        public void EnsureNBits(int n, ref BufferProcessor bufferProcessor)
         {
-            DecoderErrorCode errorCode = this.EnsureNBitsUnsafe(n, decoder);
+            DecoderErrorCode errorCode = this.EnsureNBitsUnsafe(n, ref bufferProcessor);
             errorCode.EnsureNoError();
         }
 
@@ -52,13 +51,13 @@ namespace ImageSharp.Formats.Jpg
         /// This method does not throw. Returns <see cref="DecoderErrorCode"/> instead.
         /// </summary>
         /// <param name="n">The number of bits to ensure.</param>
-        /// <param name="decoder">Jpeg decoder</param>
+        /// <param name="bufferProcessor">The <see cref="BufferProcessor"/></param>
         /// <returns>Error code</returns>
-        public DecoderErrorCode EnsureNBitsUnsafe(int n, JpegDecoderCore decoder)
+        public DecoderErrorCode EnsureNBitsUnsafe(int n, ref BufferProcessor bufferProcessor)
         {
             while (true)
             {
-                DecoderErrorCode errorCode = this.EnsureBitsStepImpl(decoder);
+                DecoderErrorCode errorCode = this.EnsureBitsStepImpl(ref bufferProcessor);
                 if (errorCode != DecoderErrorCode.NoError || this.UnreadBits >= n)
                 {
                     return errorCode;
@@ -69,34 +68,34 @@ namespace ImageSharp.Formats.Jpg
         /// <summary>
         /// Unrolled version of <see cref="EnsureNBitsUnsafe"/> for n==8
         /// </summary>
-        /// <param name="decoder">The <see cref="JpegDecoderCore"/></param>
+        /// <param name="bufferProcessor">The <see cref="BufferProcessor"/></param>
         /// <returns>A <see cref="DecoderErrorCode"/></returns>
-        public DecoderErrorCode Ensure8BitsUnsafe(JpegDecoderCore decoder)
+        public DecoderErrorCode Ensure8BitsUnsafe(ref BufferProcessor bufferProcessor)
         {
-            return this.EnsureBitsStepImpl(decoder);
+            return this.EnsureBitsStepImpl(ref bufferProcessor);
         }
 
         /// <summary>
         /// Unrolled version of <see cref="EnsureNBitsUnsafe"/> for n==1
         /// </summary>
-        /// <param name="decoder">The <see cref="JpegDecoderCore"/></param>
+        /// <param name="bufferProcessor">The <see cref="BufferProcessor"/></param>
         /// <returns>A <see cref="DecoderErrorCode"/></returns>
-        public DecoderErrorCode Ensure1BitUnsafe(JpegDecoderCore decoder)
+        public DecoderErrorCode Ensure1BitUnsafe(ref BufferProcessor bufferProcessor)
         {
-            return this.EnsureBitsStepImpl(decoder);
+            return this.EnsureBitsStepImpl(ref bufferProcessor);
         }
 
         /// <summary>
         /// Receive extend
         /// </summary>
         /// <param name="t">Byte</param>
-        /// <param name="decoder">Jpeg decoder</param>
+        /// <param name="bufferProcessor">The <see cref="BufferProcessor"/></param>
         /// <returns>Read bits value</returns>
         [MethodImpl(MethodImplOptions.AggressiveInlining)]
-        public int ReceiveExtend(int t, JpegDecoderCore decoder)
+        public int ReceiveExtend(int t, ref BufferProcessor bufferProcessor)
         {
             int x;
-            DecoderErrorCode errorCode = this.ReceiveExtendUnsafe(t, decoder, out x);
+            DecoderErrorCode errorCode = this.ReceiveExtendUnsafe(t, ref bufferProcessor, out x);
             errorCode.EnsureNoError();
             return x;
         }
@@ -105,14 +104,14 @@ namespace ImageSharp.Formats.Jpg
         /// Receive extend
         /// </summary>
         /// <param name="t">Byte</param>
-        /// <param name="decoder">Jpeg decoder</param>
+        /// <param name="bufferProcessor">The <see cref="BufferProcessor"/></param>
         /// <param name="x">Read bits value</param>
         /// <returns>The <see cref="DecoderErrorCode"/></returns>
-        public DecoderErrorCode ReceiveExtendUnsafe(int t, JpegDecoderCore decoder, out int x)
+        public DecoderErrorCode ReceiveExtendUnsafe(int t, ref BufferProcessor bufferProcessor, out int x)
         {
             if (this.UnreadBits < t)
             {
-                DecoderErrorCode errorCode = this.EnsureNBitsUnsafe(t, decoder);
+                DecoderErrorCode errorCode = this.EnsureNBitsUnsafe(t, ref bufferProcessor);
                 if (errorCode != DecoderErrorCode.NoError)
                 {
                     x = int.MaxValue;
@@ -133,10 +132,10 @@ namespace ImageSharp.Formats.Jpg
             return DecoderErrorCode.NoError;
         }
 
-        private DecoderErrorCode EnsureBitsStepImpl(JpegDecoderCore decoder)
+        private DecoderErrorCode EnsureBitsStepImpl(ref BufferProcessor bufferProcessor)
         {
             int c;
-            DecoderErrorCode errorCode = decoder.Bytes.ReadByteStuffedByteUnsafe(decoder.InputStream, out c);
+            DecoderErrorCode errorCode = bufferProcessor.Bytes.ReadByteStuffedByteUnsafe(bufferProcessor.InputStream, out c);
 
             if (errorCode != DecoderErrorCode.NoError)
             {