The main goal of this refactor was fixing #18, but - as usually - things escalated.

1. To allow EOF-s during scans (specially in progressive case), I needed to do all the block transforming after the stream decoding is done. This resulted in a strict separation of decoding steps.  It makes `JpegDecoderCore.Decode()`  as simple as:
```
this.ProcessStream(image, stream, metadataOnly);
if (!metadataOnly)
{
    this.ProcessBlockColorsIntoJpegImageChannels<TColor>();
    this.ConvertJpegPixelsToImagePixels(image);
}
```

2. To keep error management in a sane way, allowing (but not forcing!) exceptions, I introduced the following pattern for various input reader/processor methods:

```
public DecoderErrorCode ReadFullUnsafe(byte[] data, int offset, int length)
{
	......
}

public void ReadFull(byte[] data, int offset, int length)
{
    DecoderErrorCode errorCode = this.ReadFullUnsafe(data, offset, length);
    errorCode.EnsureNoError(); // Throws an exception if errorCode != DecoderErrorCode.NoError
}
```
This allows incremental refactor of error handling code from exception-based error handling to C-style spaghetti. We need to avoid exceptions  to have the same decoder performance for users with broken input images.

3. To compensate the negative performance impact of the new error-handling `if`-s I did an optimization on Huffman decoding code by converting data to `int` (almost) everywhere. This eliminates unnecessary conversions.

3. Did various refactors and code cleanups, like moving all the input reading buffering logic to a separate `InputProcessor` type.

4. The separation of scan huffman decoding (`JpegScanDecoder.DecodeBlocks()`) from the logic converting blocks into YCbCr images (`JpegScanDecoder.ProcessBlockColors()`) allowed parallel execution of `ProcessBlockColors` calls.

*An overall before vs after benchmark comparsion:*

**Before the refactor**

```
                                Method |   InputCategory |        Mean |    StdDev | Scaled | Scaled-StdDev | Allocated |
-------------------------------------- |---------------- |------------ |---------- |------- |-------------- |---------- |
     'DecodeJpegMultiple - ImageSharp' |       AllImages | 374.3229 ms | 1.2255 ms |   2.89 |          0.02 |  56.38 MB |
 'DecodeJpegMultiple - System.Drawing' |       AllImages | 129.4012 ms | 1.1507 ms |   1.00 |          0.00 |   5.78 MB |
     'DecodeJpegMultiple - ImageSharp' | LargeImagesOnly | 294.9569 ms | 0.4662 ms |   3.41 |          0.02 |  43.66 MB |
 'DecodeJpegMultiple - System.Drawing' | LargeImagesOnly |  86.5401 ms | 0.5083 ms |   1.00 |          0.00 |   5.29 MB |
     'DecodeJpegMultiple - ImageSharp' | SmallImagesOnly |  76.7843 ms | 0.4196 ms |   1.81 |          0.01 |  12.74 MB |
 'DecodeJpegMultiple - System.Drawing' | SmallImagesOnly |  42.3149 ms | 0.2948 ms |   1.00 |          0.00 | 489.48 kB |
```

**After the refactor**
```
                                Method |   InputCategory |        Mean |    StdDev | Scaled | Scaled-StdDev | Allocated |
-------------------------------------- |---------------- |------------ |---------- |------- |-------------- |---------- |
     'DecodeJpegMultiple - ImageSharp' |       AllImages | 312.7559 ms | 1.0043 ms |   2.38 |          0.01 |  56.42 MB |
 'DecodeJpegMultiple - System.Drawing' |       AllImages | 131.5888 ms | 0.4324 ms |   1.00 |          0.00 |   5.78 MB |
     'DecodeJpegMultiple - ImageSharp' | LargeImagesOnly | 252.4911 ms | 0.5561 ms |   2.83 |          0.02 |  43.67 MB |
 'DecodeJpegMultiple - System.Drawing' | LargeImagesOnly |  89.2732 ms | 0.8507 ms |   1.00 |          0.00 |   5.29 MB |
     'DecodeJpegMultiple - ImageSharp' | SmallImagesOnly |  68.0404 ms | 0.5855 ms |   1.57 |          0.02 |  12.74 MB |
 'DecodeJpegMultiple - System.Drawing' | SmallImagesOnly |  43.2593 ms | 0.4677 ms |   1.00 |          0.00 | 489.48 kB |
```

+1 `ImageSharp.Sandbox46` is now executable, and it's possible to execute BenchmarkDotNet benchmarks. This was needed to enable profiling options unavailable in test runners, like memory profiling