@JimBobSquarePants 2017-01-24
![Shia LaBeouf Slow Clap](https://cloud.githubusercontent.com/assets/385879/22240609/00988aa6-e270-11e6-88d0-d151df0e76f4.gif)

Gotta say this is some of the most impressive refactoring work i've ever seen, well done!

I've identified a small issue though. It looks like we are somehow missing the data from the first MCU. This happens in both baseline and progressive jpegs. I couldn't see an obvious cause but I'm sure you'll spot the issue straight away.

![turtle](https://cloud.githubusercontent.com/assets/385879/22240716/670ce5a2-e270-11e6-89e6-d3f737e8fed1.jpg)

We'll need to update the path in DecodeJpeg benchmark also to match new structure.