@antonfirsov 2017-01-24
I always checked my outputs, so stupid I missed this anomaly in the left corner :smile: