@antonfirsov 2017-01-28
I think I fixed it. Need to be careful with `Array.Length`on rented arrays...

@JimBobSquarePants @dlemstra @tocsoft anyone interested in a review? Or should we just Merge N'Go? :smile: