@JimBobSquarePants 2017-01-28
Great stuff! I got caught out a few times with that when deallocating the png encoder.  I'll review this sometime this evening. Gotta date by the pool this afternoon. :smile: