@JimBobSquarePants 2017-01-28
All tests are passing and all looks clean so I'm gonna merge this. Only thing I would change would be the name of `Array` struct in `DecodedBlockMemento` just to avoid clashes.