@antonfirsov 2017-01-28
FYI: the cause of codecov-io complainments is the lack of tests for the new error handler `else` branches in the Jpeg Decoder. We need more bad input images for exhaustive "EOF-tolerance" testing.